spring cleanning of Spring config
