+++ src/main/java/org/cubalibre/familink/api/config/DataSourceMySQLConfig.java
@@ -1,0 +1,44 @@
+package org.cubalibre.familink.api.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.jdbc.datasource.DriverManagerDataSource;
+
+import javax.sql.DataSource;
+
+/**
+ * Classe de type Spring JAVA Config pour créer une instance de DataSource
+ *
+ * @author DIGINAMIC
+ */
+@Configuration
+public class DataSourceMySQLConfig {
+
+    @Value("${spring.datasource.driver-class-name}")
+    private String driver;
+
+    @Value("${spring.datasource.url}")
+    private String url;
+
+    @Value("${spring.datasource.username}")
+    private String username;
+
+    @Value("${spring.datasource.password}")
+    private String password;
+
+    /**
+     * Création de la DataSource qui pointe sur la BDD MySQL spring-test
+     *
+     * @return {@link DataSource}
+     */
+    @Bean
+    public DataSource password() {
+        DriverManagerDataSource dataSource = new DriverManagerDataSource();
+        dataSource.setDriverClassName(driver);
+        dataSource.setUrl(url);
+        dataSource.setUsername(username);
+        dataSource.setPassword(password);
+        return dataSource;
+    }
+}
