+++ src/main/java/org/cubalibre/familink/api/config/JpaConfig.java
@@ -1,0 +1,52 @@
+package org.cubalibre.familink.api.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.orm.jpa.JpaTransactionManager;
+import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
+import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
+import org.springframework.transaction.PlatformTransactionManager;
+import org.springframework.transaction.annotation.EnableTransactionManagement;
+
+import javax.persistence.EntityManagerFactory;
+import javax.sql.DataSource;
+
+@Configuration
+@EnableTransactionManagement
+@EnableJpaRepositories({"org.cubalibre.familink.api.repository"})
+public class JpaConfig {
+
+    @Bean
+    public PlatformTransactionManager transactionManager(EntityManagerFactory emf) {
+        JpaTransactionManager txManager = new JpaTransactionManager();
+        txManager.setEntityManagerFactory(emf);
+        return txManager;
+    }
+
+    /**
+     * Cette configuration nécessite une source de données configurée. Elle s'utilise donc en association avec un autre
+     * fichier de configuration définissant un bean DataSource.
+     */
+    @Bean
+    public EntityManagerFactory entityManagerFactory(DataSource dataSource) {
+
+        HibernateJpaVendorAdapter vendorAdapter = new HibernateJpaVendorAdapter();
+        vendorAdapter.setGenerateDdl(true);
+        // activer les logs SQL
+        vendorAdapter.setShowSql(true);
+
+        LocalContainerEntityManagerFactoryBean factory = new LocalContainerEntityManagerFactoryBean();
+        factory.setJpaVendorAdapter(vendorAdapter);
+        // alternative au persistence.xml
+        factory.setPackagesToScan("org.cubalibre.familink.api.entite");
+        factory.setDataSource(dataSource);
+
+//		Properties jpaProperties = new Properties();
+//		jpaProperties.setProperty("javax.persistence.schema-generation.database.action", "drop-and-create");
+//		factory.setJpaProperties(jpaProperties);
+
+        factory.afterPropertiesSet();
+        return factory.getObject();
+    }
+}
