+++ src/main/java/org/cubalibre/familink/api/init/WebAppConfig.java
@@ -1,0 +1,40 @@
+package org.cubalibre.familink.api.init;
+
+import org.cubalibre.familink.api.config.DataSourceMySQLConfig;
+import org.cubalibre.familink.api.config.JpaConfig;
+import org.cubalibre.familink.api.config.ServicesConfig;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+import org.springframework.web.bind.annotation.CrossOrigin;
+import org.springframework.web.servlet.ViewResolver;
+import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
+import org.springframework.web.servlet.view.InternalResourceViewResolver;
+
+@CrossOrigin
+@Configuration
+@EnableWebMvc
+@Import({ServicesConfig.class, JpaConfig.class, DataSourceMySQLConfig.class})
+public class WebAppConfig {
+
+    @Bean
+    public ViewResolver viewResolver() {
+        return new InternalResourceViewResolver("/WEB-INF/views/", ".jsp");
+    }
+
+    @Bean
+    public WebMvcConfigurer corsConfigurer() {
+        return new WebMvcConfigurerAdapter() {
+            @Override
+            public void addCorsMappings(CorsRegistry registry) {
+                registry.addMapping("/**")
+                        .allowedOrigins("*")
+                        .allowedMethods("GET", "POST", "PUT", "PATCH", "DELETE")
+                        .allowedHeaders("Content-Type");
+            }
+        };
+    }
+}
