+++ src/main/java/org/cubalibre/familink/api/init/WebAppInitializer.java
@@ -1,47 +1,46 @@
 package org.cubalibre.familink.api.init;
 
-import javax.servlet.ServletContext;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRegistration;
-
-import org.cubalibre.familink.api.config.spring.web.WebAppConfig;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.web.WebApplicationInitializer;
 import org.springframework.web.context.ContextLoaderListener;
 import org.springframework.web.context.support.AnnotationConfigWebApplicationContext;
 import org.springframework.web.servlet.DispatcherServlet;
 
-/**
- * Classe d'implementation de l'interface fournit par Spring MVC, permetant de:
- * - charger le context Spring ( Beans )
- * - charger le context de Servlet Spring
- *
- * @author avl
- */
+import javax.servlet.ServletContext;
+import javax.servlet.ServletException;
+import javax.servlet.ServletRegistration;
+
 public class WebAppInitializer implements WebApplicationInitializer {
 
-	private static final Logger LOG = LoggerFactory.getLogger(WebAppInitializer.class.getName());
-	
-	/**
-	 * Exécutée au démarrage de l’application (Tomcat)
-	 */	
 	@Override
 	public void onStartup(ServletContext servletContext) throws ServletException {
-		LOG.info("WebAppInitializer - Demarrage Serveur .....");
-
-		// Chargement du context Spring avec des Beans, a partir d'une Configuration 
-		// (Beans - @Component : @Controller / @Service / @Repository / @Configuration)
-		AnnotationConfigWebApplicationContext webContext = new AnnotationConfigWebApplicationContext();
+        // Initialisation du contexte Spring
+        AnnotationConfigWebApplicationContext webContext = new AnnotationConfigWebApplicationContext();
 		webContext.register(WebAppConfig.class);
 
-		// Servlet Spring qui officie comme Front controller (SPRING MVC)
+		/*
+		 * <servlet> <servlet-name>dispatcher</servlet-name>
+		 * <servlet-class>org.springframework.web.servlet.DispatcherServlet</
+		 * servlet-class> <load-on-startup>1</load-on-startup> </servlet>
+		 */
 		ServletRegistration.Dynamic dispatcher = servletContext.addServlet("dispatcher", new DispatcherServlet(webContext));
-		dispatcher.setLoadOnStartup(1);
-		
-		dispatcher.addMapping("/api/*");
 
-		// Creation de contexte de Servlet (context general de Spring) ajouté par "listener" de type ServletContextListener”
+		dispatcher.setLoadOnStartup(1);
+
+		/*
+         * <servlet-mapping> <servlet-name>dispatcher</servlet-name>
+		 * <url-pattern>/mvc/*</url-pattern> </servlet-mapping>
+		 */
+        dispatcher.addMapping("/mvc/*");
+
+		/*
+		 * <listener>
+		 * 
+		 * <listener-
+		 * class>org.springframework.web.context.ContextLoaderListener</listener
+		 * -class>
+		 * 
+		 * </listener>
+		 */
 		servletContext.addListener(new ContextLoaderListener(webContext));
 	}
 }
