+++ src/main/java/org/cubalibre/familink/api/repository/ProfilRepository.java
@@ -1,0 +1,8 @@
+package org.cubalibre.familink.api.repository;
+
+import org.cubalibre.familink.api.entite.Profil;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ProfilRepository extends JpaRepository<Profil, Integer> {
+
+}
