+++ src/main/java/org/cubalibre/familink/api/services/IProfilService.java
@@ -1,0 +1,11 @@
+package org.cubalibre.familink.api.services;
+
+import org.cubalibre.familink.api.entite.Profil;
+
+import java.util.List;
+
+public interface IProfilService {
+
+    List<Profil> getProfils();
+
+}
