--- src/main/java/org/cubalibre/familink/api/services/IUserService.java
@@ -1,13 +1,0 @@
-package org.cubalibre.familink.api.services;
-
-import org.cubalibre.familink.api.entities.User;
-
-/**
- * Interface qui liste les méthodes de persistence pour la classe {@link User} que doit posséder une classe
- * d'implémentation
- *
- * @author avl
- */
-public interface IUserService {
-
-}
