--- src/main/webapp/WEB-INF/config/configuration.properties
@@ -1,22 +1,0 @@
-# Livrer independement du rest de l'application, ce fichier contient les cles-valeurs utiliser pour la configuration de l'application.
-# Les valeurs seront donner par Maven, en fonction de l'environement DEV ou PROD utilise (see profile into pom.xml)
-
-# Datasource JDBC connection
-spring.datasource.driver-class-name=@jdbc.driver@
-spring.datasource.url=@jdbc.url@
-spring.datasource.username=@jdbc.user@
-spring.datasource.password=@jdbc.pass@
-
-# Tomcat
-spring.datasource.tomcat.max-active=@jdbc.connections.max.active@
-spring.datasource.tomcat.max-idle=@jdbc.connections.max.active@
-spring.datasource.tomcat.test-on-borrow=true
-spring.datasource.tomcat.validationQuery=SELECT 1
-
-# JPA
-spring.jpa.hibernate.ddl-auto=@jpa.hibernate.ddl@
-spring.jpa.hibernate.show_sql=true
-spring.jpa.properties.hibernate.dialect = @jdbc.dialect@
-
-# Jackson
-spring.jackson.serialization.write_dates_as_timestamps=false
