--- src/main/webapp/WEB-INF/web.xml
@@ -1,7 +1,0 @@
-<!DOCTYPE web-app PUBLIC
- "-//Sun Microsystems, Inc.//DTD Web Application 2.3//EN"
- "http://java.sun.com/dtd/web-app_2_3.dtd" >
-
-<web-app>
-  	<display-name>Archetype Created Web Application</display-name>
-</web-app>
